rename to default_fracfac_generators
